+++ src/lib/binary/bytestorage.cc
@@ -148,7 +148,7 @@ ByteStorage * ByteStorage::transcode(const char * from, const char * to) {
 					size_t tmpSize = 100;
 					char * tmp = (char *) malloc(tmpSize);
 					if (tmp) {
-						size_t result = snprintf(tmp, tmpSize, "%zu", inBuf-(char *)this->data);
+						size_t result = snprintf(tmp, tmpSize, "%lu", (unsigned long) (inBuf-(char *)this->data));
 						if (result < tmpSize) { error += tmp; }
 						free(tmp);
 					}
